chore: move provider to src from common folder
